+++ src/moulti/app.py
@@ -13,11 +13,10 @@ from rich.markup import MarkupError
 from textual import work
 from textual.app import App, ComposeResult
 from textual.dom import BadIdentifier
-from textual.filter import ANSIToTruecolor
 from textual.widgets import Footer, Label
 from textual.worker import get_current_worker, NoActiveWorker
 from . import __version__ as MOULTI_VERSION
-from .ansi import replace_line_filter, AnsiThemePolicy
+from .ansi import AnsiThemePolicy
 from .protocol import PRINTABLE_MOULTI_SOCKET, clean_socket, current_instance
 from .protocol import moulti_listen, get_unix_credentials, send_json_message
 from .protocol import MoultiConnectionClosedException, MoultiProtocolException, Message, FDs
@@ -100,15 +99,11 @@ class Moulti(App):
 		steps to display the exact same thing as their terminal with no or little care for portability.
 		"""
 		ansi_behavior = os.environ.get('MOULTI_ANSI', 'verbatim')
-		if ansi_behavior == 'textual_default' or not hasattr(self, '_filters'):
+		if ansi_behavior == 'textual_default':
 			return
-		if ansi_behavior == 'verbatim':
-			# Verbatim: remove Textual's ANSIToTruecolor filter:
-			replace_line_filter(self, ANSIToTruecolor, None)
-		else: # Deal with environment variables MOULTI_ANSI and MOULTI_ANSI_THEME_*:
-			policy = AnsiThemePolicy.from_environment('MOULTI_')
-			self.logconsole(f'Applying {policy}')
-			policy.apply(self)
+		policy = AnsiThemePolicy.from_environment('MOULTI_')
+		self.logconsole(f'Applying {policy}')
+		policy.apply(self)
 		self.logconsole(f'Textual filters: {self._filters}')
 
 	def init_threads(self) -> None: